+++ src/modules/control_allocator/VehicleActuatorEffectiveness/ActuatorEffectivenessCustom.hpp
@@ -31,64 +31,29 @@
  *
  ****************************************************************************/
 
-/**
- * @file ActuatorEffectivenessTailsitterVTOL.hpp
- *
- * Actuator effectiveness for tailsitter VTOL
- */
-
 #pragma once
 
-#include "ActuatorEffectiveness.hpp"
+#include "actuator_effectiveness/ActuatorEffectiveness.hpp"
 #include "ActuatorEffectivenessRotors.hpp"
 #include "ActuatorEffectivenessControlSurfaces.hpp"
 
-#include <uORB/topics/normalized_unsigned_setpoint.h>
-
-#include <uORB/Subscription.hpp>
-
-class ActuatorEffectivenessTailsitterVTOL : public ModuleParams, public ActuatorEffectiveness
+class ActuatorEffectivenessCustom : public ModuleParams, public ActuatorEffectiveness
 {
 public:
-	ActuatorEffectivenessTailsitterVTOL(ModuleParams *parent);
-	virtual ~ActuatorEffectivenessTailsitterVTOL() = default;
+	ActuatorEffectivenessCustom(ModuleParams *parent);
+	virtual ~ActuatorEffectivenessCustom() = default;
 
 	bool getEffectivenessMatrix(Configuration &configuration, EffectivenessUpdateReason external_update) override;
-
-	int numMatrices() const override { return 2; }
-
-	void getDesiredAllocationMethod(AllocationMethod allocation_method_out[MAX_NUM_MATRICES]) const override
-	{
-		static_assert(MAX_NUM_MATRICES >= 2, "expecting at least 2 matrices");
-		allocation_method_out[0] = AllocationMethod::SEQUENTIAL_DESATURATION;
-		allocation_method_out[1] = AllocationMethod::PSEUDO_INVERSE;
-	}
-
-	void getNormalizeRPY(bool normalize[MAX_NUM_MATRICES]) const override
-	{
-		normalize[0] = true;
-		normalize[1] = false;
-	}
-
-	void allocateAuxilaryControls(const float dt, int matrix_index, ActuatorVector &actuator_sp) override;
 
 	void updateSetpoint(const matrix::Vector<float, NUM_AXES> &control_sp, int matrix_index,
 			    ActuatorVector &actuator_sp, const matrix::Vector<float, NUM_ACTUATORS> &actuator_min,
 			    const matrix::Vector<float, NUM_ACTUATORS> &actuator_max) override;
 
-
-	void setFlightPhase(const FlightPhase &flight_phase) override;
-
-	const char *name() const override { return "VTOL Tailsitter"; }
+	const char *name() const override { return "Custom"; }
 
 protected:
-	ActuatorEffectivenessRotors _mc_rotors;
-	ActuatorEffectivenessControlSurfaces _control_surfaces;
+	ActuatorEffectivenessRotors _motors;
+	ActuatorEffectivenessControlSurfaces _torque;
 
-	uint32_t _forwards_motors_mask{};
-
-	int _first_control_surface_idx{0}; ///< applies to matrix 1
-
-	uORB::Subscription _flaps_setpoint_sub{ORB_ID(flaps_setpoint)};
-	uORB::Subscription _spoilers_setpoint_sub{ORB_ID(spoilers_setpoint)};
+	uint32_t _motors_mask{};
 };
