+++ src/modules/control_allocator/VehicleActuatorEffectiveness/ActuatorEffectivenessFixedWing.hpp
@@ -33,21 +33,36 @@
 
 #pragma once
 
-#include "ActuatorEffectiveness.hpp"
+#include "actuator_effectiveness/ActuatorEffectiveness.hpp"
+#include "ActuatorEffectivenessRotors.hpp"
+#include "ActuatorEffectivenessControlSurfaces.hpp"
 
-class ActuatorEffectivenessRoverAckermann : public ActuatorEffectiveness
+#include <uORB/topics/normalized_unsigned_setpoint.h>
+
+class ActuatorEffectivenessFixedWing : public ModuleParams, public ActuatorEffectiveness
 {
 public:
-	ActuatorEffectivenessRoverAckermann() = default;
-	virtual ~ActuatorEffectivenessRoverAckermann() = default;
+	ActuatorEffectivenessFixedWing(ModuleParams *parent);
+	virtual ~ActuatorEffectivenessFixedWing() = default;
 
 	bool getEffectivenessMatrix(Configuration &configuration, EffectivenessUpdateReason external_update) override;
+
+	const char *name() const override { return "Fixed Wing"; }
+
+	void allocateAuxilaryControls(const float dt, int matrix_index, ActuatorVector &actuator_sp) override;
 
 	void updateSetpoint(const matrix::Vector<float, NUM_AXES> &control_sp, int matrix_index,
 			    ActuatorVector &actuator_sp, const matrix::Vector<float, NUM_ACTUATORS> &actuator_min,
 			    const matrix::Vector<float, NUM_ACTUATORS> &actuator_max) override;
 
-	const char *name() const override { return "Rover (Ackermann)"; }
 private:
-	uint32_t _motors_mask{};
+	ActuatorEffectivenessRotors _rotors;
+	ActuatorEffectivenessControlSurfaces _control_surfaces;
+
+	uORB::Subscription _flaps_setpoint_sub{ORB_ID(flaps_setpoint)};
+	uORB::Subscription _spoilers_setpoint_sub{ORB_ID(spoilers_setpoint)};
+
+	int _first_control_surface_idx{0}; ///< applies to matrix 1
+
+	uint32_t _forwards_motors_mask{};
 };
