+++ src/modules/control_allocator/VehicleActuatorEffectiveness/ActuatorEffectivenessRoverAckermann.hpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2023 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2021 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -33,35 +33,21 @@
 
 #pragma once
 
-#include "ActuatorEffectiveness.hpp"
-#include "ActuatorEffectivenessRotors.hpp"
+#include "actuator_effectiveness/ActuatorEffectiveness.hpp"
 
-class ActuatorEffectivenessUUV : public ModuleParams, public ActuatorEffectiveness
+class ActuatorEffectivenessRoverAckermann : public ActuatorEffectiveness
 {
 public:
-	ActuatorEffectivenessUUV(ModuleParams *parent);
-	virtual ~ActuatorEffectivenessUUV() = default;
+	ActuatorEffectivenessRoverAckermann() = default;
+	virtual ~ActuatorEffectivenessRoverAckermann() = default;
 
 	bool getEffectivenessMatrix(Configuration &configuration, EffectivenessUpdateReason external_update) override;
-
-	void getDesiredAllocationMethod(AllocationMethod allocation_method_out[MAX_NUM_MATRICES]) const override
-	{
-		allocation_method_out[0] = AllocationMethod::SEQUENTIAL_DESATURATION;
-	}
-
-	void getNormalizeRPY(bool normalize[MAX_NUM_MATRICES]) const override
-	{
-		normalize[0] = true;
-	}
 
 	void updateSetpoint(const matrix::Vector<float, NUM_AXES> &control_sp, int matrix_index,
 			    ActuatorVector &actuator_sp, const matrix::Vector<float, NUM_ACTUATORS> &actuator_min,
 			    const matrix::Vector<float, NUM_ACTUATORS> &actuator_max) override;
 
-	const char *name() const override { return "UUV"; }
-
-protected:
-	ActuatorEffectivenessRotors _rotors;
-
+	const char *name() const override { return "Rover (Ackermann)"; }
+private:
 	uint32_t _motors_mask{};
 };
